# openapi.yaml
openapi: 3.1.0
info:
  title: 'Saturn Platform'
  version: '0.1'
servers:
  -
    url: 'https://app.example.com/api/v1'
    description: 'Saturn Platform Cloud API. Change the host to your own instance if you are self-hosting.'
paths:
...
  /applications/public:
    post:
      tags:
        - Applications
      summary: 'Create (Public)'
      description: 'Create new application based on a public git repository.'
      operationId: create-public-application
      requestBody:
        description: 'Application object that needs to be created.'
        required: true
        content:
          application/json:
            schema:
              required:
                - project_uuid
                - server_uuid
                - environment_name
                - environment_uuid
                - git_repository
                - git_branch
                - build_pack
                - ports_exposes
              properties:
                project_uuid:
                  description: 'The project UUID.'
                  type: string
                server_uuid:
                  description: 'The server UUID.'
                  type: string
                environment_name:
                  description: 'The environment name.'
                  type: string
                environment_uuid:
                  description: 'The environment UUID.'
                  type: string
                git_repository:
                  description: 'The git repository URL.'
                  type: string
                git_branch:
                  description: 'The git branch.'
                  type: string
                build_pack:
                  description: 'The build pack type.'
                  type: string
                  enum: [nixpacks, static, dockerfile, dockercompose]
                ports_exposes:
                  description: 'The ports to expose.'
                  type: string
                instant_deploy:
                  description: 'Deploy instantly after creation.'
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Application created successfully.'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '409':
          description: 'Domain conflicts detected.'
      security:
        -
          bearerAuth: []
  /applications/private-github-app:
    post:
      tags:
        - Applications
      summary: 'Create (Private - GH App)'
      description: 'Create new application based on a private repository through a Github App.'
      operationId: create-private-github-app-application
      requestBody:
        description: 'Application object that needs to be created.'
        required: true
        content:
          application/json:
            schema:
              required:
                - project_uuid
                - server_uuid
                - environment_name
                - environment_uuid
                - github_app_uuid
                - git_repository
                - git_branch
                - build_pack
                - ports_exposes
              properties:
                project_uuid:
                  description: 'The project UUID.'
                  type: string
                server_uuid:
                  description: 'The server UUID.'
                  type: string
                environment_name:
                  description: 'The environment name.'
                  type: string
                environment_uuid:
                  description: 'The environment UUID.'
                  type: string
                github_app_uuid:
                  description: 'The Github App UUID.'
                  type: string
                git_repository:
                  description: 'The git repository URL.'
                  type: string
                git_branch:
                  description: 'The git branch.'
                  type: string
                build_pack:
                  description: 'The build pack type.'
                  type: string
                  enum: [nixpacks, static, dockerfile, dockercompose]
                ports_exposes:
                  description: 'The ports to expose.'
                  type: string
                instant_deploy:
                  description: 'Deploy instantly after creation.'
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Application created successfully.'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '409':
          description: 'Domain conflicts detected.'
      security:
        -
          bearerAuth: []
  /applications/private-deploy-key:
    post:
      tags:
        - Applications
      summary: 'Create (Private - Deploy Key)'
      description: 'Create new application based on a private repository through a Deploy Key.'
      operationId: create-private-deploy-key-application
      requestBody:
        description: 'Application object that needs to be created.'
        required: true
        content:
          application/json:
            schema:
              required:
                - project_uuid
                - server_uuid
                - environment_name
                - environment_uuid
                - private_key_uuid
                - git_repository
                - git_branch
                - build_pack
                - ports_exposes
              properties:
                project_uuid:
                  description: 'The project UUID.'
                  type: string
                server_uuid:
                  description: 'The server UUID.'
                  type: string
                environment_name:
                  description: 'The environment name.'
                  type: string
                environment_uuid:
                  description: 'The environment UUID.'
                  type: string
                private_key_uuid:
                  description: 'The private key UUID.'
                  type: string
                git_repository:
                  description: 'The git repository URL.'
                  type: string
                git_branch:
                  description: 'The git branch.'
                  type: string
                build_pack:
                  description: 'The build pack type.'
                  type: string
                  enum: [nixpacks, static, dockerfile, dockercompose]
                ports_exposes:
                  description: 'The ports to expose.'
                  type: string
                instant_deploy:
                  description: 'Deploy instantly after creation.'
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Application created successfully.'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '409':
          description: 'Domain conflicts detected.'
      security:
        -
          bearerAuth: []
  /applications/dockerfile:
    post:
      tags:
        - Applications
      summary: 'Create (Dockerfile)'
      description: 'Create new application based on a simple Dockerfile.'
      operationId: create-dockerfile-application
      requestBody:
        description: 'Application object that needs to be created.'
        required: true
        content:
          application/json:
            schema:
              required:
                - project_uuid
                - server_uuid
                - environment_name
                - environment_uuid
                - dockerfile
              properties:
                project_uuid:
                  description: 'The project UUID.'
                  type: string
                server_uuid:
                  description: 'The server UUID.'
                  type: string
                environment_name:
                  description: 'The environment name.'
                  type: string
                environment_uuid:
                  description: 'The environment UUID.'
                  type: string
                dockerfile:
                  description: 'The Dockerfile content (base64 encoded).'
                  type: string
                instant_deploy:
                  description: 'Deploy instantly after creation.'
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Application created successfully.'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '409':
          description: 'Domain conflicts detected.'
      security:
        -
          bearerAuth: []
  /applications/dockerimage:
    post:
      tags:
        - Applications
      summary: 'Create (Docker Image)'
      description: 'Create new application based on a prebuilt docker image'
      operationId: create-dockerimage-application
      requestBody:
        description: 'Application object that needs to be created.'
        required: true
        content:
          application/json:
            schema:
              required:
                - project_uuid
                - server_uuid
                - environment_name
                - environment_uuid
                - docker_registry_image_name
                - ports_exposes
              properties:
                project_uuid:
                  description: 'The project UUID.'
                  type: string
                server_uuid:
                  description: 'The server UUID.'
                  type: string
                environment_name:
                  description: 'The environment name.'
                  type: string
                environment_uuid:
                  description: 'The environment UUID.'
                  type: string
                docker_registry_image_name:
                  description: 'The docker registry image name.'
                  type: string
                docker_registry_image_tag:
                  description: 'The docker registry image tag.'
                  type: string
                ports_exposes:
                  description: 'The ports to expose.'
                  type: string
                instant_deploy:
                  description: 'Deploy instantly after creation.'
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Application created successfully.'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '409':
          description: 'Domain conflicts detected.'
      security:
        -
          bearerAuth: []
  /applications/dockercompose:
    post:
      tags:
        - Applications
      summary: 'Create (Docker Compose)'
      description: 'Create new application based on a docker-compose file.'
      operationId: create-dockercompose-application
      requestBody:
        description: 'Application object that needs to be created.'
        required: true
        content:
          application/json:
            schema:
              required:
                - project_uuid
                - server_uuid
                - environment_name
                - environment_uuid
                - docker_compose_raw
              properties:
                project_uuid:
                  description: 'The project UUID.'
                  type: string
                server_uuid:
                  description: 'The server UUID.'
                  type: string
                environment_name:
                  description: 'The environment name.'
                  type: string
                environment_uuid:
                  description: 'The environment UUID.'
                  type: string
                docker_compose_raw:
                  description: 'The Docker Compose raw content (base64 encoded).'
                  type: string
                instant_deploy:
                  description: 'Deploy instantly after creation.'
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Application created successfully.'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '409':
          description: 'Domain conflicts detected.'
      security:
        -
          bearerAuth: []
...
components:
...
  responses:
    '400':
      description: 'Invalid token.'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Invalid token.'
            type: object
    '401':
      description: Unauthenticated.
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: Unauthenticated.
            type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Go to `Keys & Tokens` / `API tokens` and create a new token. Use the token as the bearer token.'
      scheme: bearer
...